+++ libs/utils/plotting.py
@@ -16,7 +16,7 @@ from intellistop import VFStopsResultType
 
 from .formatting import dates_extractor_list
 from .plot_utils import (
-    candlesticks, utils, dual_plotting, generic
+    bar_charting, candlesticks, utils, dual_plotting, generic, speciality
 )
 
 
@@ -24,12 +24,16 @@ class PlotType(Enum):
     CANDLESTICKS = 'candlesticks'
     DUAL_PLOTTING = 'dual_plotting'
     GENERIC_PLOTTING = 'generic_plotting'
+    BAR_CHART = 'bar_chart'
+    SPECIALITY = 'speciality'
 
 
 FUNCTIONS = {
     PlotType.CANDLESTICKS: candlesticks.candlestick_plot,
     PlotType.DUAL_PLOTTING: dual_plotting.dual_plotting,
-    PlotType.GENERIC_PLOTTING: generic.generic_plotting
+    PlotType.GENERIC_PLOTTING: generic.generic_plotting,
+    PlotType.BAR_CHART: bar_charting.bar_chart,
+    PlotType.SPECIALITY: speciality.specialty_plotting,
 }
 
 
@@ -43,307 +47,6 @@ def generate_plot(plot_type: PlotType, fund: pd.DataFrame, **kwargs):
     # view = kwargs.get('view', '')
     kwargs['save_fig'] = not kwargs.get('plot_output', False)
     FUNCTIONS.get(plot_type, {})(fund, **kwargs)
-
-
-def generic_plotting(list_of_plots: list, **kwargs):
-    """Generic Plotting
-
-    Plot any number of plots together, using same Y-axis. Note, each plot in list_of_plots can be a
-    different length as long as the "x" key is utilized and that the corresponding x-plot has the
-    same length as the corresponding y-plot. Great use for a fund's price versus various moving
-    average plots.
-
-    Arguments:
-        list_of_plots {list} -- list of y-value data sets to be plotted (multiple), a list of lists.
-                                If only one used, ensure the y-list is wrapped in outer [].
-
-    Optional Args:
-        x {list} -- x-value data, can be list of lists, length(s) must match length(s) of y plot(s).
-                    NOTE: for list of list of dates, ensure each list is wrapped in a list()
-                    (default: {[]})
-        colors {list} -- color data for each plot (default: {[]})
-        title {str} -- title of plot (default: {''})
-        legend {list} -- list of plot labels in legend (default: {[]})
-        save_fig {bool} -- True will save as 'filename' (default: {False})
-        filename {str} -- path to save plot (default: {'temp_generic_plot.png'})
-        ylabel {str} -- label for y axis (default: {''})
-
-    Returns:
-        None
-    """
-    # pylint: disable=too-many-branches,too-many-statements
-    register_matplotlib_converters()
-
-    x_list = kwargs.get('x', [])
-    colors = kwargs.get('colors', [])
-    title = kwargs.get('title', '')
-    legend = kwargs.get('legend', [])
-    save_fig = kwargs.get('save_fig', False)
-    filename = kwargs.get('filename', 'temp_generic_plot')
-    ylabel = kwargs.get('ylabel', '')
-
-    if len(colors) > 0:
-        if len(colors) != len(list_of_plots):
-            print(
-                f"{utils.WARNING}Warning: lengths of plots ({len(list_of_plots)}) and colors " +
-                f"({len(colors)}) do not match in generic_plotting.{utils.NORMAL}")
-            return None
-
-    if len(x_list) > 0:
-        if isinstance(x_list[0], (list, pd.core.indexes.datetimes.DatetimeIndex)):
-            for i, plot in enumerate(list_of_plots):
-                if len(plot) != len(x_list[i]):
-                    print(
-                        f"{utils.WARNING}Warning: lengths of plots ({len(plot)}) and x " +
-                        f"({len(x_list[i])}) " +
-                        f"do not match in generic_plotting.{utils.NORMAL}")
-                    return None
-
-    fig, axis = plt.subplots()
-
-    if len(x_list) < 1:
-        x_list = dates_extractor_list(list_of_plots[0])
-        for i, fig in enumerate(list_of_plots):
-            if len(colors) > 0:
-                plt.plot(x_list, fig, colors[i])
-            else:
-                plt.plot(x_list, fig)
-
-    else:
-        if isinstance(x_list[0], (list, pd.core.indexes.datetimes.DatetimeIndex)):
-            for i, fig_y in enumerate(list_of_plots):
-                if len(colors) > 0:
-                    plt.plot(x_list[i], fig_y, colors[i])
-                else:
-                    plt.plot(x_list[i], fig_y)
-
-        else:
-            for i, fig_y in enumerate(list_of_plots):
-                if len(colors) > 0:
-                    plt.plot(x_list, fig_y, colors[i])
-                else:
-                    plt.plot(x_list, fig_y)
-
-    plt.title(title)
-    if len(legend) > 0:
-        plt.legend(legend)
-    if ylabel != '':
-        plt.ylabel(ylabel)
-
-    utils.plot_xaxis_disperse(axis)
-
-    try:
-        if save_fig:
-            temp_path = os.path.join("output", "temp")
-            if not os.path.exists(temp_path):
-                # For functions, this directory may not exist.
-                plt.close(fig)
-                plt.clf()
-                return None
-
-            filename = os.path.join(temp_path, filename)
-            if os.path.exists(filename):
-                os.remove(filename)
-
-            plt.savefig(filename)
-
-        else:
-            plt.show()
-
-    except: # pylint: disable=bare-except
-        print(
-            f"{utils.WARNING}Warning: plot failed to render in 'generic_plotting' of title: " +
-            f"{title}{utils.NORMAL}")
-
-    plt.close('all')
-    plt.clf()
-    return None
-
-
-def bar_chart(data: list, **kwargs):
-    """Bar Chart
-
-    Exclusively used for MACD, On Balance Volume, Awesome Oscillator
-
-    Arguments:
-        data {list} -- list of y-values to be plotted
-
-    Optional Args:
-        x {list} -- x-value data (default: {[]}) (length of lists)
-        position {pd.DataFrame} -- fund data, plotted as normal plot (default: {[]})
-        title {str} -- title of plot (default: {''})
-        save_fig {bool} -- True will save as 'filename' (default: {False})
-        filename {str} -- path to save plot (default: {'temp_bar_chart.png'})
-        positive {bool} -- True plots all color bars positively (default: {False})
-        bar_delta {bool} -- True: y[i] > y[i-1] -> green, else red (default: {False})
-
-    Returns:
-        None
-    """
-    # pylint: disable=too-many-locals,too-many-branches,too-many-statements
-    register_matplotlib_converters()
-
-    x_list = kwargs.get('x', [])
-    position = kwargs.get('position', [])
-    title = kwargs.get('title', '')
-    save_fig = kwargs.get('save_fig', False)
-    filename = kwargs.get('filename', 'temp_bar_chart.png')
-    all_positive = kwargs.get('all_positive', False)
-    bar_delta = kwargs.get('bar_delta', False)
-
-    if len(x_list) < 1:
-        x_list = list(range(len(data)))
-
-    colors = []
-    positive = []
-    bar_awesome = []
-
-    for i, bar_data in enumerate(data):
-        if bar_data > 0.0:
-            positive.append(bar_data)
-            colors.append('green')
-        elif bar_data < 0.0:
-            positive.append(-1.0 * bar_data)
-            colors.append('red')
-        else:
-            positive.append(bar_data)
-            colors.append('black')
-
-        if i == 0:
-            bar_awesome.append(colors[-1])
-        else:
-            if bar_data < data[i-1]:
-                bar_awesome.append('red')
-            elif bar_data > data[i-1]:
-                bar_awesome.append('green')
-            else:
-                bar_awesome.append(bar_awesome[-1])
-
-    if all_positive:
-        data = positive
-
-    fig, ax1 = plt.subplots()
-    if bar_delta:
-        barlist = ax1.bar(x_list, data, width=1, color=bar_awesome)
-    else:
-        barlist = ax1.bar(x_list, data, width=1, color=colors)
-
-    if not bar_delta:
-        for i in range(1, len(data)):
-            if data[i] > 0.0:
-                if data[i] < data[i-1]:
-                    barlist[i].set_alpha(0.3)
-            else:
-                if data[i] > data[i-1]:
-                    barlist[i].set_alpha(0.3)
-
-    plt.title(title)
-
-    if len(position) > 0:
-        ax2 = ax1.twinx()
-        ax2.plot(position['Close'])
-
-    utils.plot_xaxis_disperse(ax1)
-
-    try:
-        if save_fig:
-            temp_path = os.path.join("output", "temp")
-            if not os.path.exists(temp_path):
-                # For functions, this directory may not exist.
-                plt.close(fig)
-                plt.clf()
-                return
-
-            filename = os.path.join(temp_path, filename)
-            if os.path.exists(filename):
-                os.remove(filename)
-            plt.savefig(filename)
-
-        else:
-            plt.show()
-
-    except: # pylint: disable=bare-except
-        print(
-            f"{utils.WARNING}Warning: plot failed to render in 'bar_chart' of name: {title}{utils.NORMAL}")
-
-    plt.close('all')
-    plt.clf()
-
-
-def specialty_plotting(list_of_plots: list, **kwargs):
-    """Specialty Plotting
-
-    Plot various datasets against others (different y-axes). Similar to dual plotting, in a sense.
-
-    Arguments:
-        list_of_plots {list} -- list of y-value datasets to be plotted (multiple)
-
-    Optional Args:
-        x {list} -- x-value data (default: {[]}) (length of lists)
-        alt_ax_index {list} -- list of other y-value datasets to be plotted on other axis
-                                (default: {[]})
-        title {str} -- title of plot (default: {''})
-        legend {list} -- list of plot labels in legend (default: {[]})
-        save_fig {bool} -- True will save as 'filename' (default: {False})
-        filename {str} -- path to save plot (default: {'temp_specialty_plot.png'})
-
-    Returns:
-        None
-    """
-    register_matplotlib_converters()
-
-    x_list = kwargs.get('x', [])
-    alt_ax_index = kwargs.get('alt_ax_index', [])
-    title = kwargs.get('title', '')
-    legend = kwargs.get('legend', [])
-    save_fig = kwargs.get('save_fig', False)
-    filename = kwargs.get('filename', 'temp_specialty_plot.png')
-
-    if len(x_list) < 1:
-        x_list = dates_extractor_list(list_of_plots[0])
-    fig, axis = plt.subplots()
-
-    for i, plot_item in enumerate(list_of_plots):
-        if i not in alt_ax_index:
-            axis.plot(x_list, plot_item)
-
-    ax2 = axis.twinx()
-    for i, plot_item in enumerate(list_of_plots):
-        if i in alt_ax_index:
-            ax2.plot(x_list, plot_item, color='tab:purple')
-
-    plt.title(title)
-    ax2.set_ylabel(legend[0])
-    if len(legend) > 0:
-        plt.legend(legend)
-
-    utils.plot_xaxis_disperse(axis)
-
-    try:
-        if save_fig:
-            temp_path = os.path.join("output", "temp")
-            if not os.path.exists(temp_path):
-                # For functions, this directory may not exist.
-                plt.close(fig)
-                plt.clf()
-                return
-
-            filename = os.path.join(temp_path, filename)
-            if os.path.exists(filename):
-                os.remove(filename)
-
-            plt.savefig(filename)
-
-        else:
-            plt.show()
-
-    except: # pylint: disable=bare-except
-        print(
-            f"{utils.WARNING}Warning: plot failed to render in 'specialty plotting' " +
-            f"of title: {title}{utils.NORMAL}")
-
-    plt.close('all')
-    plt.clf()
 
 
 def shape_plotting(main_plot: pd.DataFrame, **kwargs):
